Triple Andybiotic (#403)
* Enable cart-style COP cost logic

* modifyAbilityStars() should scale with fatigue

* Link star costs for everyone but Ave and Brenner

* Add some helpful(?) comments re: modifyAbilityPower/Stars()

* Add base class for AW2/3 COs

* TrilogyAbility gets jank cost logic

* Add AW2 Andy

* Fix iteration bounds for CO picker GUI

* Enable showing charge bars with more flexible scales

* AW3 star value -> int

* Add AWDS Andy

* Fix wording on mechanics blurbs

* Un-bug my supposedly not bugged implementation

* Add initial AWBW Andy impl

* Implement AWBW healing jank

* The exception's exception was considered a bug and fixed. Praise be to Walker

* Compilation fixes

* Kill off TODO

* The label for current energy in the game stats bugged me

* Boiled pasta

* More explicit variable names, per review

* Clarify modifyAbilityPower() comment

* Image optimization pass?

* Clarify TrilogyCommander comments

* pixelsPerPowerPlanck

* Blurbify, TrilogyCO -> BuggedChargeCO

* BuggedCommander->AW2And3 base

* Delete ability.adjustCost()

+++ src/Units/UnitModel.java
@@ -69,7 +69,7 @@ public abstract class UnitModel implements Serializable, ITargetable, UnitModLis
   public static final int MAXIMUM_HEALTH = 100;
   public String name;
   public long role;
-  public double abilityPowerValue;
+  public int abilityPowerValue; // In percent of a star's value per GUI HP
   public int maxAmmo;
   public int maxFuel;
   public int fuelBurnIdle;
@@ -86,7 +86,7 @@ public abstract class UnitModel implements Serializable, ITargetable, UnitModLis
   public Set<TerrainType> unloadExclusionTerrain = new HashSet<TerrainType>();
 
   public UnitModel(String pName, long pRole, int cost, int pAmmoMax, int pFuelMax, int pIdleFuelBurn, int pVision, int pMovePower,
-      MoveType pPropulsion, UnitActionFactory[] actions, WeaponModel[] pWeapons, double powerValue)
+      MoveType pPropulsion, UnitActionFactory[] actions, WeaponModel[] pWeapons, int powerValue)
   {
     this(pName, pRole, cost, pAmmoMax, pFuelMax, pIdleFuelBurn, pVision, pMovePower, pPropulsion, powerValue);
 
@@ -101,7 +101,7 @@ public abstract class UnitModel implements Serializable, ITargetable, UnitModLis
   }
 
   public UnitModel(String pName, long pRole, int cost, int pAmmoMax, int pFuelMax, int pIdleFuelBurn, int pVision, int pMovePower,
-      MoveType pPropulsion, ArrayList<UnitActionFactory> actions, ArrayList<WeaponModel> pWeapons, double powerValue)
+      MoveType pPropulsion, ArrayList<UnitActionFactory> actions, ArrayList<WeaponModel> pWeapons, int powerValue)
   {
     this(pName, pRole, cost, pAmmoMax, pFuelMax, pIdleFuelBurn, pVision, pMovePower, pPropulsion, powerValue);
     baseActions.addAll(actions);
@@ -109,7 +109,7 @@ public abstract class UnitModel implements Serializable, ITargetable, UnitModLis
   }
 
   private UnitModel(String pName, long pRole, int cost, int pAmmoMax, int pFuelMax, int pIdleFuelBurn, int pVision, int pMovePower,
-      MoveType pPropulsion, double powerValue)
+      MoveType pPropulsion, int powerValue)
   {
     name = pName;
     role = pRole;
